server/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
python-docx
